Move PR approval and unapproval assertions to `BorsTester`
